+++ src/base/b-virtual-scroll/test/runners/functional/render-next.js
@@ -28,6 +28,30 @@ module.exports = (page) => {
 		getContainerChildCount = () => component.evaluate((ctx) => ctx.$refs.container.childElementCount);
 
 	beforeEach(async () => {
+		await page.evaluate(() => {
+			globalThis.removeCreatedComponents();
+
+			const baseAttrs = {
+				theme: 'demo',
+				option: 'section',
+				optionProps: ({current}) => ({'data-index': current.i})
+			};
+
+			const scheme = [
+				{
+					attrs: {
+						...baseAttrs,
+						id: 'target'
+					}
+				}
+			];
+
+			globalThis.renderComponents('b-virtual-scroll', scheme);
+		});
+
+		await h.bom.waitForIdleCallback(page);
+		await h.component.waitForComponentStatus(page, '.b-virtual-scroll', 'ready');
+
 		node = await h.dom.waitForEl(page, '#target');
 		component = await h.component.waitForComponent(page, '#target');
 		container = await h.dom.waitForRef(node, 'container');