+++ src/core/dom/index.ts
@@ -13,6 +13,10 @@
 
 import { deprecate } from 'core/functools';
 
+//#if runtime has core/dom/in-view
+export * from 'core/dom/in-view';
+//#endif
+
 /**
  * Wraps the specified function as an event handler with delegation.
  * This function can be used as a decorator or as a simple function.