+++ src/form/b-button/test/index.js
@@ -6,10 +6,29 @@
  * https://github.com/V4Fire/Client/blob/master/LICENSE
  */
 
+const
+	h = include('tests/helpers');
+
 module.exports = async (page, {browserType, tmpDir}) => {
 	await page.screenshot({path: `${tmpDir}/example-${browserType}.png`});
 
 	describe('b-button', () => {
+		beforeAll(async () => {
+			await page.evaluate(() => {
+				globalThis.removeCreatedComponents();
+
+				const scheme = [
+					{
+						attrs: {
+							id: 'target'
+						}
+					}
+				];
+
+				globalThis.renderComponents('b-button', scheme);
+			});
+		});
+
 		it('getting a component instance from a DOM node', async () => {
 			const
 				component = await (await page.$('.b-button')).getProperty('component'),