+++ src/pages/p-v4-components-demo/index.js
@@ -7,8 +7,8 @@
  */
 
 const
-	components = (require('config').build.components || []).map(({name}) => name);
+	components = (require('config').build.components ?? []).map(({name}) => name);
 
 package('p-v4-components-demo')
 	.extends('i-static-page')
-	.dependencies('b-v4-component-demo', components);
+	.dependencies('b-v4-component-demo', 'b-virtual-scroll', 'b-button', 'b-router', components);