+++ src/super/i-block/modules/dom/interface.ts
@@ -11,3 +11,28 @@ import iBlock from 'super/i-block/i-block';
 export interface ElCb<CTX extends iBlock = iBlock> {
 	(this: CTX, el: Element): any;
 }
+
+export const
+	inViewInstanceStoreSymbol: unique symbol = Symbol.for('in-view instance store');
+
+/**
+ * @see https://github.com/microsoft/TypeScript/issues/1863
+ */
+export const
+	inViewInstanceStore = <any>inViewInstanceStoreSymbol;
+
+export interface DOMManipulationOptions {
+	/**
+	 * If true and the source node has a component property,
+	 * then when the destructor is called, $destroy of the component will be called too.
+	 */
+	destroyIfComponent?: boolean;
+
+	/**
+	 * Async group
+	 *
+	 * @see [[Async]]
+	 * @default `asyncComponents`
+	 */
+	group?: string;
+}